This patch provides the write capability. The user can now issue insert / update and delete statements for the foreign tables using the mysql FDW.
This patch has also pushed-down the foreign table where clause
to the foreign server. The where condition on the foreign table
Will be executed on the foreign server hence there will be fewer rows to
to bring across to PG. This is a performance feature.

This patch also implements column push-down and only
brings back the columns that are part of the select target list.