Error (#75) Introduce new server option "use_remote_estimate".
A new foreign server option "use_remote_estimate" is
added. If this option is true for a server then server gets the
number of rows from remote MySQL server. Server issues
an "EXPLAIN" call for the query to remote MySQL server and
gets the rows column and filtered column. Using the rows
and filtered column, it calculates the actual number of rows for
the query.