Change INFO on established mysql connection to DEBUG1
The user should not be bothered with it in normal operation because it
interferes with the output of queries executed. Moreover, INFO is
not possible to suppress using client_min_messages for having
reproducible regression test results (the message contains the mysql
server version). To debug, users can "SET client_min_messages = debug;"
to get the server connection information.