Qualify table name with the database name while identifying a unique column.
In mysql_is_column_unique() to identifying a unique column from the
table, we use an EXPLAIN query.  But the table name in that must be
qualified with the database name otherwise we may refer to the table
from the earlier connected database.  And it that is different, then
we may end up accessing the wrong table. Fix that.

On the way, if there is no database name given in the FOREIGN TABLE
options, then like we do in mysql_deparse_relation(), use the schema
name of the foreign table as the database name.

FDW-126, Suraj Kharage, reviewed by Vaibhav Dalvi, further revised by
Jeevan Chalke.