Fix crash with ANALYZE on a non-existing remote table.
To get the table size in ANALYZE operation, we run a SELECT query on
the MySQL database by passing the database name and table name.  The
query returns no rows if the table does not present in the remote
database.  However, we have not checked for this possibility and thus
we ended up with the server crash later.  Throw an error if we get no
rows from the remote query execution.

Reported by Suraj Kharage while working on FDW-126.

Suraj Kharage, reviewed, and a few cosmetic changes by Jeevan Chalke.