Fix UPDATE to account BR trigger changes.
Earlier, only columns that are present in the UPDATE query got
updated at the MySQL side.  However, in any column value which is
not present in the UPDATE query but modified in the BEFORE ROW
trigger was not updated at MySQL side.  Fix that by marking all
columns as updatable in such cases.  The fix for this was given
by Francois Payette through pull request #194 on GitHub, which is
further revised by Suraj Kharage.

Also, updating a row-identifier column is not supported.  However,
this was broken when the BR update trigger is trying to update that.
Fix that by comparing old and new value for the row-identifier column
and throwing an error if they are not the same.

Reported on GitHub through issues #193 by Francois Payette.

FDW-193, Suraj Kharage, reviewed by Vaibhav Dalvi and me.