Fix crash with IMPORT FOREIGN SCHEMA when executed repeatedly
mysqlImportForeignSchema() internally implements IMPORT FOREIGN
SCHEMA, while doing so it needs to know of server options and hence
it calls mysql_get_options().  mysql_get_options() in a PG_TRY block
wrongly uses server oid to retrieve the foreign table by passing it
to GetForeignTable() which throws an error and increases the depth of
stack errordata.  This error is caught in a PG_CATCH block, and now
the same oid is used to retrieve the foreign server without clearing
the error state or re-throwing it.  Hence, when this repeatedly
executed, it exceeds the max ERRORDATA_STACK_SIZE defined as 5.  Fix
this by passing a flag to know if we have a server oid or a table oid
and accordingly call either GetForeignServer() or GetForeignTable().

Reported on GitHub as issue #210 by Emmanuel (man1100).

FDW-248, Suraj Kharage, initial analysis and review by Vaibhav Dalvi.