Fix compiler warning due to clash with mysql library macro
mysql_fdw uses a default port number when the 'port' server option
is not provided. This was defined as MYSQL_PORT, but this used to
clash with the mysql client library header definitions in some mysql
versions. Fix this by renaming the macro to avoid the warning.

FDW-286, Suraj Kharage, reviewed by Jeevan Ladhe.