Adjust some tests to throw generic error messages using plpgsql
Some of the tests throw errors that vary depending on which mysql
server version, or which platform they are being run on. Due to this
the regression tests fail as the errors or output did not match with
that in the expected output files. Fix them by catching as generic
exceptions in plpgsql blocks and then throwing generic error
messages. This will ensure that the errors are uniform across
platform or different versions of mysql server.

FDW-203, Suraj Kharage, reviewed by Jeevan Ladhe.