Fix parameterized query when use_remote_estimates is enabled.
For estimating the number of rows on the remote server for a given
query, mysql_fdw sends an explain query to the MySQL server. If the
query is parameterized, the parameters are positionally marked with
'?'(question marks), which MySQL does not understand and throws an
error. Instead, fix this by sending the parameter list as dummy null
while estimating the rows.

FDW-400, Suraj Kharage, reviewed by Vaibhav Dalvi, reported by
Rajkumar Raghuwanshi.