Add reconnect option to control automatic reconnection to MySQL server.
The MySQL client library can perform an automatic reconnection to the
server if it finds the connection is down when one attempts to send
a statement to the server to be executed, and it is controlled by
MySQL option MYSQL_OPT_RECONNECT. This allows a seamless operation in
case the connection got dropped intermittently. Add a mysql_fdw
server-level option 'reconnect' to control this behavior for
mysql_fdw internal connections.

Reported on GitHub through issue #215 by Magmatrix.

FDW-350, Suraj Kharage, reviewed by Jeevan Ladhe.