Do not push expressions having mutable functions to the remote server.
If an expression that includes a mutable function is sent on the
remote server via remote query its result may not be stable. Instead,
check if the expression has a mutable function, if yes do not send
that expression to the remote MySQL server.

Reported on GitHub as issue #226 by user robins.

FDW-411, Suraj Kharage, reviewed by Vaibhav Dalvi.