Push down aggregates to remote MySQL servers.
Push aggregates to the remote MySQL server instead of fetching all
of the rows and aggregating them locally. This gives a very good
performance boost for the cases where aggregates can be pushed down.

The push-down is currently limited to aggregate functions min, max,
sum, avg, and count, to avoid pushing down the functions that are not
present on the MySQL server. Also, aggregate filters and orders are
not pushed down as MySQL does not support them.

FDW-132, Suraj Kharage, reviewed by Vaibhav Dalvi, Jeevan Ladhe and
tested by Rajkumar Raghuwanshi.