Fix text column data truncation when reading very long data.
If data in the MySQL table is greater than 64k, then the column data
is truncated when it is read from the foreign table. This is because
the column is allocated fixed size of MAXDATALEN which is defined as
64k. Fix this by setting STMT_ATTR_UPDATE_MAX_LENGTH to retrieve
max_length and use it to allocate the column.

FDW-156, Suraj Kharage, reviewed by Vaibhav Dalvi, tested by
Kashif Zeeshan.