Push down ORDER BY to remote MySQL servers.
If possible, consider a path that adds the ORDER BY clause to the
remote SQL so that we get the ordered result set from the foreign
server itself.  It might help us to have an efficient merge join.

NULLs behavior is opposite on the MySQL server.  Thus to get an
equivalent result, we add the "expression IS NULL" clause at the
beginning of each of the ORDER BY expressions.

FDW-130, Suraj Kharage, reviewed by Vaibhav Dalvi and Jeevan Chalke,
tested by Kashif Zeeshan.