Add support for the TRUNCATE command to truncate foreign tables.
RESTART IDENTITY and CONTINUE IDENTITY modifiers are completely
ignored, and only TRUNCATE <table_name> is sent to the MySQL server.
And thus, MySQL's default behavior is observed irrespective of these
modifiers, i.e., sequence, if any, will be reset.  If this behavior
is NOT expected, then mark the table as not truncatable.  Also,
CASCADE is not supported; RESTRICT is fine as it is a default
behavior anyway.

FDW-519, Suraj Kharage, reviewed by Sravan Velagandula and
Jeevan Chalke, tested by Kashif Zeeshan.