Push down the "LIMIT n" clause when OFFSET is NULL.
We were rejecting this case for no reason.  Since OFFSET NULL does
nothing and is the same as omitting it, we can simply push the
"LIMIT n" part to the remote.  This is a small improvement in the
LIMIT/OFFSET pushdown feature.

FDW-527, Suraj Kharage, reviewed by Sravan Velagandula, tested by
Kashif Zeeshan.