Add configuration file approach to control the pushdowns.
In WHERE clause pushdown, all the built-in functions and operators
were pushed down to the remote server.  This results in an error when
that function or operator is not present on the remote server.  We
have avoided this issue in join clauses (join pushdown) and aggregates
(aggregate pushdown) by sending only the known set of functions and
operators to the remote server.

To have more control over these pushdowns, add an infrastructure of
the configuration file where we can define the list of operators and
functions which can be pushed down to the remote server safely.  This
configuration file is read and stored in an in-memory hash table to
check the pushability.

Users can easily add or modify the list per the requirements and must
be in the given format.  Also, user can check the in-memory listing
by calling the function mysql_fdw_display_pushdown_list().

FDW-322, Suraj Kharage, reviewed by Vaibhav Dalvi, further reviewed
and a few changes by Jeevan Chalke, tested by Ajay Pal.